+++ packages/sdk-common/src/utils/address.test.ts
@@ -1,6 +1,6 @@
-import { normalizeAptosAddress } from "./address"
+import { normalizeAptosAddress, normalizeAptosNetwork } from "./address"
 
-describe("address utilities", () => {
+describe("address/network utilities", () => {
   test("normalize address #1", () => {
     const normalizedAddress = normalizeAptosAddress("0x001")
     expect(normalizedAddress).toBe("0x0000000000000000000000000000000000000000000000000000000000000001")
@@ -9,5 +9,15 @@ describe("address utilities", () => {
   test("normalize address #2", () => {
     const normalizedAddress = normalizeAptosAddress("0x99999999999999999")
     expect(normalizedAddress).toBe("0x0000000000000000000000000000000000000000000000099999999999999999")
+  })
+
+  test("normalize network #1", () => {
+    const normalizedAddress = normalizeAptosNetwork("mainnet")
+    expect(normalizedAddress).toBe("Mainnet")
+  })
+
+  test("normalize network #2", () => {
+    const normalizedAddress = normalizeAptosNetwork("Testnet")
+    expect(normalizedAddress).toBe("Testnet")
   })
 })