+++ packages/sdk/src/sdk-blockchains/ethereum/fill.ts
@@ -18,6 +18,7 @@ import { Platform } from "@rarible/api-client"
 import type { AmmTradeInfo } from "@rarible/ethereum-api-client"
 import { Warning } from "@rarible/logger/build"
 import { extractBlockchain } from "@rarible/sdk-common"
+import type { GetOrderBuyTxRequest, TransactionData } from "@rarible/protocol-ethereum-sdk/build/order/fill-order/types"
 import type {
   BatchFillRequest,
   FillRequest,
@@ -69,6 +70,7 @@ export class EthereumFill {
     this.acceptBidBasic = this.acceptBidBasic.bind(this)
     this.batchBuyBasic = this.batchBuyBasic.bind(this)
     this.getBuyAmmInfo = this.getBuyAmmInfo.bind(this)
+    this.getBuyTxData = this.getBuyTxData.bind(this)
   }
 
   async buyBasic(request: BuySimplifiedRequest): Promise<IBlockchainTransaction> {
@@ -507,6 +509,10 @@ export class EthereumFill {
       hash: request.hash,
       numNFTs: request.numNFTs,
     })
+  }
+
+  async getBuyTxData(request: GetOrderBuyTxRequest): Promise<TransactionData> {
+    return this.sdk.order.getBuyTxData(request)
   }
 
   async batchBuyBasic(